+++ src/hooks/useLinkTwitter.ts
@@ -1,49 +1,40 @@
-import { useMemo } from 'react'
 import { useAsyncFn } from 'react-use'
 import urlcat from 'urlcat'
 import { useAccount, useSignMessage } from 'wagmi'
 import { FIREFLY_API_ROOT } from '../constants/api'
 import { fetchJSON } from '../helpers/fetchJSON'
 import { useAccountStore } from '../store/accountStore'
-import { LoginResponse, TwitterAuthorizeResponse } from '../types/api'
+import { TwitterAuthorizeResponse } from '../types/api'
 import { useHandleError } from './useHandleError'
+import { useLogin } from './useLogin'
 import { useToast } from './useToast'
 
 export function useLinkTwitter() {
   const account = useAccount()
   const toast = useToast()
   const { signMessageAsync } = useSignMessage()
-  const message = useMemo(() => `Link X ${Date.now()}`, [])
   const handleError = useHandleError()
 
   const { updateToken } = useAccountStore()
+  const { token } = useAccountStore()
+  const login = useLogin()
 
   return useAsyncFn(async () => {
     if (!account.address) return
     try {
-      const signed = await signMessageAsync({ message })
-
-      const loginUrl = urlcat(FIREFLY_API_ROOT, '/v1/mask_stake/wallet/login')
-      const loginRes = await fetchJSON<LoginResponse>(loginUrl, {
-        method: 'POST',
-        body: JSON.stringify({
-          original_message: message,
-          signature_message: signed.slice(2), // omit 0x
-          wallet_address: account.address,
-        }),
-      })
-      if (loginRes.code === 200) {
-        updateToken(loginRes.data.token)
-      } else {
-        throw new Error('Failed to login')
+      let jwtToken = token
+      if (!jwtToken) {
+        jwtToken = await login.mutateAsync()
       }
 
       const url = urlcat(FIREFLY_API_ROOT, '/v1/mask_stake/twitter/authorize', {
-        original_message: message,
-        signature_message: signed.slice(2), // omit 0x
         wallet_address: account.address,
       })
-      const res = await fetchJSON<TwitterAuthorizeResponse>(url)
+      const res = await fetchJSON<TwitterAuthorizeResponse>(url, {
+        headers: {
+          Authorization: `Bearer ${jwtToken}`,
+        },
+      })
       if (res.code !== 200) {
         console.error('Failed to get twitter authorize', res.message, res.reason)
         toast({
@@ -58,5 +49,5 @@ export function useLinkTwitter() {
       if (handleError(err)) return
       throw err
     }
-  }, [account.address, signMessageAsync, updateToken])
+  }, [account.address, signMessageAsync, login.mutateAsync, updateToken])
 }
